requirements.txt
cryptography==41.0.3
requests==2.31.0
finnhub-python==2.4.20
flask==2.3.3
flask-socketio==5.3.6
gunicorn==21.2.0
python-socketio==5.8.0
eventlet==0.33.3
